requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pyautogen==0.2.0
streamlit==1.28.2
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1 